pythonBackend/requirements.txt
praisonaiagents>=0.0.156
praisonai>=2.2.82
praisonai-tools>=0.0.22
openai>=1.98.0
pillow>=10.0.0
requests>=2.32.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6